// contracts/typescript/models/DayInput.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues, parseDate, parseDateTime, serializeDate, serializeDateTime } from '../runtime';
import type { Availability } from './Availability';
import {
    AvailabilityFromJSON,
    AvailabilityFromJSONTyped,
    AvailabilityToJSON,
    AvailabilityToJSONTyped,
} from './Availability';
import type { WorkLocation } from './WorkLocation';
import {
    WorkLocationFromJSON,
    WorkLocationFromJSONTyped,
    WorkLocationToJSON,
    WorkLocationToJSONTyped,
} from './WorkLocation';

/**
 *
 * @export
 * @interface DayInput
 */
export interface DayInput {
    /**
     *
     */
    availability: Availability;
    /**
     *
     */
    baseDayId?: string | null;
    /**
     *
     */
    basePlanVersion?: number | null;
    /**
     *
     */
    cancel?: boolean;
    /**
     *
     */
    localDate: Date;
    /**
     *
     */
    location: WorkLocation;
}



/**
 * Check if a given object implements the DayInput interface.
 */
export function instanceOfDayInput(value: object): value is DayInput {
    if (!('availability' in value) || value['availability'] === undefined) return false;
    if (!('localDate' in value) || value['localDate'] === undefined) return false;
    if (!('location' in value) || value['location'] === undefined) return false;
    return true;
}

export function DayInputFromJSON(json: any): DayInput {
    return DayInputFromJSONTyped(json, false);
}

export function DayInputFromJSONTyped(json: any, ignoreDiscriminator: boolean): DayInput {
    if (json == null) {
        return json;
    }
    return {

        'availability': AvailabilityFromJSON(json['availability']),
        'baseDayId': json['baseDayId'] === undefined ? undefined : json['baseDayId'] === null ? null : json['baseDayId'],
        'basePlanVersion': json['basePlanVersion'] === undefined ? undefined : json['basePlanVersion'] === null ? null : json['basePlanVersion'],
        'cancel': json['cancel'] == null ? undefined : json['cancel'],
        'localDate': (json['localDate'] == null ? json['localDate'] : parseDate(json['localDate'])),
        'location': WorkLocationFromJSON(json['location']),
    };
}

export function DayInputToJSON(json: any): DayInput {
    return DayInputToJSONTyped(json, false);
}

export function DayInputToJSONTyped(value?: DayInput | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {

        'availability': AvailabilityToJSON(value['availability']),
        'baseDayId': value['baseDayId'],
        'basePlanVersion': value['basePlanVersion'],
        'cancel': value['cancel'],
        'localDate': value['localDate'] == null ? value['localDate'] : serializeDate(value['localDate']),
        'location': WorkLocationToJSON(value['location']),
    };
}