// contracts/typescript/models/WorkLocation.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 */
export const WorkLocation = {
    Unplanned: 'Unplanned',
    OfficeSwitzerland: 'OfficeSwitzerland',
    RemotePortugal: 'RemotePortugal',
} as const;
export type WorkLocation = typeof WorkLocation[keyof typeof WorkLocation];


export function instanceOfWorkLocation(value: any): boolean {
    for (const key in WorkLocation) {
        if (Object.prototype.hasOwnProperty.call(WorkLocation, key)) {
            if (WorkLocation[key as keyof typeof WorkLocation] === value) {
                return true;
            }
        }
    }
    return false;
}

export function WorkLocationFromJSON(json: any): WorkLocation {
    return WorkLocationFromJSONTyped(json, false);
}

export function WorkLocationFromJSONTyped(json: any, ignoreDiscriminator: boolean): WorkLocation {
    return json as WorkLocation;
}

export function WorkLocationToJSON(value?: WorkLocation | null): any {
    return value as any;
}

export function WorkLocationToJSONTyped(value: any, ignoreDiscriminator: boolean): WorkLocation {
    return value as WorkLocation;
}