// contracts/typescript/models/Availability.ts
/* tslint:disable */
/* eslint-disable */
/**
 * HomeOfficeReservation API
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: v1
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 *
 * @export
 */
export const Availability = {
    Working: 'Working',
    Leave: 'Leave',
    Unavailable: 'Unavailable',
} as const;
export type Availability = typeof Availability[keyof typeof Availability];


export function instanceOfAvailability(value: any): boolean {
    for (const key in Availability) {
        if (Object.prototype.hasOwnProperty.call(Availability, key)) {
            if (Availability[key as keyof typeof Availability] === value) {
                return true;
            }
        }
    }
    return false;
}

export function AvailabilityFromJSON(json: any): Availability {
    return AvailabilityFromJSONTyped(json, false);
}

export function AvailabilityFromJSONTyped(json: any, ignoreDiscriminator: boolean): Availability {
    return json as Availability;
}

export function AvailabilityToJSON(value?: Availability | null): any {
    return value as any;
}

export function AvailabilityToJSONTyped(value: any, ignoreDiscriminator: boolean): Availability {
    return value as Availability;
}